third push
I decided to not move on with my original sketch because I found that it was very difficult to get the code to work the way I wanted it to. In my code, I commented out some of the struggles and previous unsuccessful attempts. This is why I have a new composition, which progress is also included.